requirements.txt
feedparser==6.0.11
requests==2.31.0
google-generativeai==0.8.4
edge-tts==6.1.10
pydub==0.25.1
resend==2.23.0
python-telegram-bot==21.0
pyyaml==6.0.1
jinja2==3.1.3
beautifulsoup4==4.12.3
aiohttp==3.9.3
pytz==2024.1
python-Levenshtein==0.25.1
python-dotenv==1.0.1
